requirements.txt
diffusers==0.33.1
transformers==4.51.3
accelerate==1.6.0
peft==0.15.2
ftfy==6.3.1
torch>=2.0.0
torchvision>=0.15.0
pillow>=9.0.0
opencv-python>=4.5.0 